style cleanup
