--- SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/interfaces/buffer_management/BufferingOperation.java
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) 2018 The University of Manchester
- */
-package uk.ac.manchester.spinnaker.front_end.interfaces.buffer_management;
-
-/**
- *
- * @author Christian-B
- */
-public enum BufferingOperation {
-    BUFFER_READ(0),
-    BUFFER_WRITE(1);
-
-    public final int value;
-
-    BufferingOperation(int value) {
-        this.value = value;
-    }
-    
-}
