+++ SpiNNaker-front-end/src/main/java/uk/ac/manchester/spinnaker/front_end/interfaces/buffer_management/storage_objects/ChannelBufferState.java
@@ -4,9 +4,10 @@
 package uk.ac.manchester.spinnaker.front_end.interfaces.buffer_management.storage_objects;
 
 import java.nio.ByteBuffer;
-import uk.ac.manchester.spinnaker.front_end.interfaces.buffer_management.BufferingOperation;
 
 /**
+ * Stores information related to a single channel output buffering state,
+ *     as it is retrieved at the end of a simulation on the SpiNNaker system.
  *
  * @see <a href=
  * "https://github.com/SpiNNakerManchester/SpiNNFrontEndCommon/blob/master/spinn_front_end_common/interface/buffer_management/storage_objects/channel_buffer_state.py">
@@ -15,31 +16,31 @@ import uk.ac.manchester.spinnaker.front_end.interfaces.buffer_management.Bufferi
  */
 public class ChannelBufferState {
 
-    // start buffering area memory address (32 bits)
+    /** start buffering area memory address (32 bits). */
     public final int startAddress;
 
-    // address where data was last written (32 bits)
+    /** address where data was last written (32 bits). */
     public final int  currentWrite;
 
-    // address where the DMA write got up to (32 bits)
+    /** address where the DMA write got up to (32 bits). */
     public final int currentDmaWrite;
 
-    // address where data was last read (32 bits)
-    public int currentRead;
+    /** address where data was last read (32 bits). */
+    private int currentRead;
 
-    // The address of first byte after the buffer (32 bits)
+    /** The address of first byte after the buffer (32 bits). */
     public final int endAddress;
 
-    // The ID of the region (8 bits)
+    /** The ID of the region (8 bits). */
     public final byte regionId;
 
-    // # True if the region overflowed during the simulation (8 bits)
+    /** True if the region overflowed during the simulation (8 bits). */
     public final byte missingInfo;
 
-    // # Last operation performed on the buffer - read or write (8 bits)
+    /** Last operation performed on the buffer - read or write (8 bits). */
     private BufferingOperation lastBufferOperation;
 
-    // bool check for if its extracted data from machine
+    /** bool check for if its extracted data from machine. */
     private boolean updateCompleted;
 
     /**
@@ -51,6 +52,11 @@ public class ChannelBufferState {
      */
     public static final int STATE_SIZE = 24;
 
+    /**
+     * Creates a new ChannelBufferState reading from the ByteBuffer.
+     *
+     * @param data ByteBuffer assumed to hold channel buffer state.
+     */
     public ChannelBufferState(ByteBuffer data) {
         // _CHANNEL_BUFFER_PATTERN = struct.Struct("<IIIIIBBBx")
         startAddress = data.getInt();
